#49 Resolviendo hash para routing, habilitando ** para solución de autenticación, cambio de llamado de webcomponent en app component
