+++ src/app/app.component.html
@@ -1,1 +1,4 @@
-<router-outlet></router-outlet>
+<ng-uui-oas [environment]="environment"></ng-uui-oas>
+<div *ngIf="loadRouting">
+    <router-outlet></router-outlet>
+</div>
