+++ src/app/app.module.ts
@@ -1,5 +1,5 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -24,10 +24,11 @@ import { ToasterModule } from 'angular2-toaster';
     BrowserAnimationsModule,
 
   ],
+  schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
   providers: [
     AppService,
-    { provide: APP_BASE_HREF, useValue: '/' },
-    //{ provide: LocationStrategy, useClass: HashLocationStrategy }
+    //{ provide: APP_BASE_HREF, useValue: '/' },
+    { provide: LocationStrategy, useClass: HashLocationStrategy }
   ],
 
   bootstrap: [AppComponent]
